[#388] Drive typed-quantity bypass constructors out of the astrodyn gateway
Move kernels and their typed siblings out of the workspace-root `astrodyn`
gateway (`src/**`) into the appropriate physics crates so the gateway becomes
a curated re-export surface, and widen `scripts/check_no_escape_hatches.sh`
to scan `src/**` (skipping `#[cfg(test)]` modules).

Relocations:
- Gravity accumulation (`accumulate_gravity`, `accumulate_relativistic_corrections`,
  their typed siblings, `evaluate_body_gravity_typed`, `run_gravity_stage`,
  `ResolvedSource`, `ResolvedRelativisticSource`, `GravityBodyInputs`) →
  `astrodyn_gravity::accumulate`.
- Force collection (`collect_and_resolve_forces` and its typed sibling) →
  `astrodyn_interactions::forces`. Promoted `astrodyn_dynamics` from dev- to
  runtime-dep of `astrodyn_interactions` (no cycle: the reverse edge does
  not exist).
- Derived-state typed lifts (`compute_body_solar_beta`,
  `compute_body_solar_beta_typed`, `compute_body_geodetic`,
  `compute_body_geodetic_typed`) → `astrodyn_math::{solar_beta,geodetic}`.
- Frame orchestration (`frame_origin`, `frame_origin_typed`,
  `compute_relative_state_typed`, `sync_pfix_rotation`) →
  `astrodyn_frames::orchestration`. `evaluate_and_apply_frame_switch` and
  `FrameSwitchTargetMissing` stay in the gateway (they consume
  gateway-only `FrameSwitchConfig` / `SwitchSense`).

Deletions:
- Gateway thin wrappers `compute_drag` / `compute_gravity_torque` /
  `compute_cannonball_srp` (and their typed siblings) — physics-crate
  typed siblings already existed and are now re-exported. Bevy adapter
  and `astrodyn_runner` consumers now compute the small `t_inertial_struct`
  / `t_parent_this` matrix preamble themselves.
- `VehicleBuilder::with_state(s: TranslationalState)` — the only untyped
  builder convenience. Production scenarios (`mars_orbit`, `clementine_lunar`,
  `mercury`, `apollo`) construct typed `TranslationalStateTyped<RootInertial>`
  directly via `Vec3Ext::m_at` / `m_per_s_at`. Test fixtures and examples
  call the typed `with_translational` with `from_untyped_unchecked` (still
  in sanctioned scopes).
- Untyped `from_pos_vel_with_mu` raw-DVec3 entry; recipe presets construct
  typed `Position<PlanetInertial<P>>` / `Velocity<PlanetInertial<P>>` via
  `Vec3Ext` directly.

Architectural exception (locked-in hybrid decision): `integrate_body_typed`
(`src/integration.rs`) stays in the gateway with `// allowed:` annotations.
The integrator orchestration composes gateway-owned types
(`crate::integrator::IntegratorType` wrapper, `FlatPlateState`,
`IntegrableObject`) and surfaces Bevy/runner-aware diagnostic messages, so
relocating it would either strip the API insulation layer or pull those
wrappers into `astrodyn_dynamics`. The kernel itself is the documented
orchestration boundary; three `from_raw_si` / `from_untyped_unchecked`
sites are annotated inline with the reason.

Lint changes:
- `scripts/check_no_escape_hatches.sh` now scans `src/**` in addition to
  `crates/astrodyn_bevy/src/**`. The awk script tracks `#[cfg(test)] mod
  ... { ... }` brace depth so test fixtures are skipped — they're a
  sanctioned boundary.
- Header comment and failure message updated to reflect the new scope.

Verification: `cargo fmt --check`, `cargo clippy --workspace --tests
--examples -- -D warnings`, `cargo test --workspace -- --skip tier3_`
(1303 passed, 0 failed), `cargo test --test invariant_coverage`, and
`bash scripts/check_no_escape_hatches.sh` are all green.
