+++ crates/astrodyn_interactions/Cargo.toml
@@ -15,11 +15,11 @@ all-features = true
 [dependencies]
 glam = { workspace = true }
 astrodyn_atmosphere = { path = "../astrodyn_atmosphere", version = "0.1.0" }
+astrodyn_dynamics = { path = "../astrodyn_dynamics", version = "0.1.0" }
 astrodyn_quantities = { path = "../astrodyn_quantities", version = "0.1.0" }
 uom = { workspace = true }
 
 [dev-dependencies]
-astrodyn_dynamics = { path = "../astrodyn_dynamics", version = "0.1.0" }
 astrodyn_ephemeris = { path = "../astrodyn_ephemeris", version = "0.1.0" }
 astrodyn_gravity = { path = "../astrodyn_gravity", version = "0.1.0" }
 astrodyn_math = { path = "../astrodyn_math", version = "0.1.0" }
