+++ crates/astrodyn_verif_jeod/examples/earth_moon.rs
@@ -115,7 +115,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         velocity: INIT_VEL,
     };
     let clementine = VehicleBuilder::new()
-        .with_state(trans)
+        .with_translational(astrodyn::TranslationalStateTyped::<
+            astrodyn_quantities::frame::RootInertial,
+        >::from_untyped_unchecked(&trans))
         .three_dof_point_mass(vehicle::clementine_mass())
         .rk4()
         .gravity(GravityControl::new_nonspherical(moon, 60, 60, false))
